pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kuru-mm-py"
version = "0.1.3"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiocache>=0.12.3",
    "aiohttp>=3.9.0",
    "asyncio>=4.0.0",
    "loguru>=0.7.3",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.0.0",
    "web3>=7.14.0",
    "websockets>=15.0.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.setuptools.package-data]
"src.abis" = ["*.json"]
